first

+++ controllers/categories/update.php
@@ -17,32 +17,32 @@ if (isset($_POST['update'])) {
     } else if (strlen($categorytitle) > 20) {
         $_SESSION['validation-errors'][] = 'title must be less than 21 char';
     }
-    //check if upload button is clicked or not
-    if (isset($_FILES['image']['name'])) {
+    //check if image is selected or not
+    if (isset($_FILES['image']['name']) && !empty($_FILES['image']['name'])) {
         $image_name = $_FILES['image']['name'];
         $image_type = $_FILES['image']['type'];
         $image_tmp = $_FILES['image']['tmp_name'];
         $image_error = $_FILES['image']['error'];
         $image_size = $_FILES['image']['size'];
-        $image_extension = pathinfo($image_name, PATHINFO_EXTENSION);
-        $image_new_name = rand() . '.' . $image_extension;
-        $data = move_uploaded_file($image_tmp, '../../images/categories/' .$image_new_name);
-        //    var_dump($data);
-        //    die();
-
-        // remove the old image
-        if (!empty($current_image)) {
-            //image is available
-            $remove_path = "../../images/categories/" .$current_image;
+        if (!empty($image_name)) {
+            //upload new image
+            $image_extension = pathinfo($image_name, PATHINFO_EXTENSION);
+            $image_name = rand() . '.' . $image_extension;
+            $data = move_uploaded_file($image_tmp, '../../images/categories/'.$image_name);
+            // remove the old image
+            //     //image is available
+            $remove_path = "../../images/categories/" . $current_image;
             $remove = unlink($remove_path);
-            if ($remove == false) {
-                $_SESSION['error'] = "Image Failed To Be Removed.";
-                header("location:" . SITEURL . "adminpanel/category-manage.php");
-            } else {
-                // upload button not clicked
-                $image_name = $current_image;
-            }
+              if ($remove == false) {
+            $_SESSION['error'] = "Image Failed To Be Removed.";
+            header("location:" . SITEURL . "adminpanel/category-manage.php");
         }
+        }else {
+            $image_name = $current_image;
+        }
+    }else {
+        // upload button not clicked
+        $image_name = $current_image;
     }
 
     //check empty errors
@@ -50,9 +50,9 @@ if (isset($_POST['update'])) {
         header('location:' . SITEURL . 'adminpanel/update_category.php');
     } else
         //5- create the sql query to be updated
-        $sql = "UPDATE categories SET title = '$categorytitle', featured = '$featured', active = '$active', `image` = '$image_new_name' WHERE id = '$id'";
-    //     echo $sql;
-    //    die();
+        $sql = "UPDATE categories SET title = '$categorytitle', featured = '$featured', active = '$active', `image` = '$image_name' WHERE id = '$id'";
+    // echo $sql;
+    // die();
     //execute the query
     $result = mysqli_query($connection, $sql);
     // var_dump($result);
